requirements.txt
discord.py>=1.7.3
Pillow>=8.3.0
jishaku>=1.18
git+https://github.com/Rapptz/asqlite
git+https://github.com/Rapptz/discord-ext-menus
numpy>=1.18
lark-parser>=0.12